[ Feat ] add stepper to sign-in page
